Move import file permissions class eval to behavior
